Fix bug comparing strings instead of numbers, plus some quality of life improvements.
