+++ webapp/components/creactivistas/actus/form/FormQuestion.js
@@ -2,8 +2,7 @@ import React from 'react';
 import Grid from '@material-ui/core/Grid';
 import Slider from '@material-ui/core/Slider';
 
-import { marks } from '../constants'
-
+import { marks } from '../../../../lib/actus/constants'
 
 class FormQuestion extends React.Component {
   constructor(props) {
